+++ components/molecules/LocaleSwitcherSelect.tsx
@@ -1,0 +1,66 @@
+'use client';
+
+import { usePathname, useRouter } from '@/i18n/navigation';
+import { useParams } from 'next/navigation';
+import { ChangeEvent, ReactNode, useTransition } from 'react';
+
+type Props = {
+  children: ReactNode;
+  defaultValue: string;
+  label: string;
+};
+
+export default function LocaleSwitcherSelect({ children, defaultValue, label }: Props) {
+  const router = useRouter();
+  const [isPending, startTransition] = useTransition();
+  const pathname = usePathname();
+  const params = useParams();
+
+  function onSelectChange(event: ChangeEvent<HTMLSelectElement>) {
+    const nextLocale = event.target.value;
+    startTransition(() => {
+      router.replace(
+        // @ts-expect-error -- TypeScript will validate that only known `params`
+        // are used in combination with a given `pathname`. Since the two will
+        // always match for the current route, we can skip runtime checks.
+        { pathname, params },
+        { locale: nextLocale },
+      );
+    });
+  }
+
+  return (
+    <label className="relative text-eleva-neutral-900 dark:text-eleva-neutral-100">
+      <p className="sr-only">{label}</p>
+      <select
+        className="appearance-none bg-transparent py-2 pl-2 pr-8"
+        defaultValue={defaultValue}
+        disabled={isPending}
+        onChange={onSelectChange}
+      >
+        {children}
+      </select>
+      <span className="pointer-events-none absolute right-2 top-1/2 -translate-y-1/2">
+        {isPending ? (
+          <span className="inline-block h-4 w-4 animate-spin rounded-full border-2 border-t-transparent" />
+        ) : (
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="16"
+            height="16"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            aria-hidden="true"
+          >
+            <title>Dropdown Arrow</title>
+            <path d="m6 9 6 6 6-6" />
+          </svg>
+        )}
+      </span>
+    </label>
+  );
+}
